+++ hari_client/__init__.py
@@ -1,7 +1,8 @@
 from hari_client.client import errors
 from hari_client.client.client import HARIClient
 from hari_client.config.config import Config
+from hari_client.config.config import HARIUploaderConfig
 from hari_client.models import models
 from hari_client.upload import hari_uploader
 
-__all__ = [HARIClient, Config, errors, models, hari_uploader]
+__all__ = [HARIClient, Config, errors, models, hari_uploader, HARIUploaderConfig]