+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 ### Fixes
 
 - correct typo in development installation guidelines [PR#43](https://github.com/quality-match/hari-client/pull/43)
+- `query` argument of multiple methods is now serialized properly to fit the backend's implementation of a query parameter array [PR#49](https://github.com/quality-match/hari-client/pull/49)
+  - get_medias
+  - get_media_count
+  - get_media_objects
+  - get_media_object_count
+  - get_attributes
+  - get_attribute_metadata
 
 ### Internal
 
@@ -21,6 +28,7 @@
 ## [3.0.0] - 06.12.2024
 
 ### New Features
+
 - added cant solve ratio to all attribute models. [PR#47](https://github.com/quality-match/hari-client/pull/47)
 
 ### Breaking Changes